requirements.txt
django>=4.0,<5.0

djangorestframework

celery[redis]

psycopg2-binary
